+++ .github/workflows/feature.yml
@@ -8,9 +8,31 @@ on:
 
 permissions:
   contents: read
+  packages: read
+
+permissions:
+  contents: read
+  packages: read
 
 jobs:
-  testing-and-building:
+  testing:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up JDK 17
+        uses: actions/setup-java@v3
+        with:
+          java-version: '17'
+          distribution: 'temurin'
+      - name: Grant execute permissions for gradlew
+        run: chmod +x gradlew
+        ##- name: Run unit tests
+        ##run: ./gradlew test
+        ##- name: Run integration test
+        ##run: ./gradlew integrationTest
+
+  building:
+    needs: testing
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
@@ -23,21 +45,14 @@ jobs:
         run: chmod +x gradlew
       - name: Build with Gradle
         run: ./gradlew build
-      - name: Run Unit Tests
-        run: ./gradlew test
-      - name: Run Integration Tests
-        run: ./gradlew integrationTest
-      - name: Run System Tests
-        run: ./gradlew systemTest
       - name: Upload JAR as artifact
         uses: actions/upload-artifact@v3
         with:
           name: app-jar
-          path: build/libs/DevOps-ExaminationProject-0.0.1-SNAPSHOT.jar
-
+          path: build/libs/final-0.0.1-SNAPSHOT.jar
 
   publish:
-    needs: testing-and-building
+    needs: building
     permissions: write-all
     runs-on: ubuntu-latest
     steps:
@@ -68,8 +83,34 @@ jobs:
         run: |
           docker push ghcr.io/${{ env.REPO_NAME }}:development-latest
 
-  deploy:
+  system-testing:
     needs: publish
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+      - name: Set up JDK 17
+        uses: actions/setup-java@v3
+        with:
+          java-version: '17'
+          distribution: 'temurin'
+      - name: Grant execute permissions for gradlew
+        run: chmod +x gradlew
+      - name: Run system-testing
+        run: ./gradlew systemTest
+    services:
+      app:
+        image: ghcr.io/${{ env.REPO_NAME }}:development-latest
+        credentials:
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+        ports:
+          - 8080:8080
+        ##options: --health-cmd="curl -f http://localhost:8080" --health-interval=10s --health-timeout=5s --health-retries=3
+
+
+  deploy:
+    needs: system-testing
     runs-on: ubuntu-latest
     steps:
       - name: Checkout code
@@ -77,4 +118,4 @@ jobs:
       - name: Convert repository name to lowercase
         run: echo "REPO_NAME=$(echo ${{ github.repository }} | tr '[:upper:]' '[:lower:]')" >> $GITHUB_ENV
       - name: Deploy Fake
-        run: docker run -d "ghcr.io/${{ env.REPO_NAME }}:development-latest"
+        run: echo "Deploying ghcr.io/${{ env.REPO_NAME }}:development-latest"
