+++ src/integrationTest/java/com/example/DevOpsExaminationProject/CarControllerTest.java
@@ -1,18 +1,26 @@
 package com.example.DevOpsExaminationProject;
 
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.web.servlet.MockMvc;
 
+import static io.restassured.RestAssured.given;
+import static org.hamcrest.Matchers.notNullValue;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@SpringBootTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
-class GreetingControllerIntegrationTest {
+class CarControllerTest {
+
+    @LocalServerPort
+    private int port;
 
     @Autowired
     private MockMvc mockMvc;
@@ -22,5 +30,23 @@ class GreetingControllerIntegrationTest {
         mockMvc.perform(get("/greet"))
                 .andExpect(status().isOk())
                 .andExpect(content().string("Hello, World!"));
+    }
+
+    @BeforeEach
+    void setUp(){
+        RestAssured.baseURI = "http://localhost";
+        RestAssured.port = port;
+    }
+
+    @Test
+    void getCar() {
+        given()
+                .when()
+                .get("/car")
+                .then()
+                .statusCode(200)
+                .assertThat()
+                .body("light", notNullValue())
+                .body("engine", notNullValue());
     }
 }
