+++ src/main/java/com/example/DevOpsExaminationProject/CarService.java
@@ -1,0 +1,10 @@
+package com.example.DevOpsExaminationProject;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CarService {
+   public String greet(){
+       return "Hello, World!";
+   }
+}
