+++ src/main/java/com/example/DevOpsExaminationProject/controllers/CarController.java
@@ -1,11 +1,17 @@
 package com.example.DevOpsExaminationProject.controllers;
 
 import com.example.DevOpsExaminationProject.Car;
+import com.example.DevOpsExaminationProject.CarService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class CarController {
+
+    @Autowired
+    CarService carService;
 
     @GetMapping("/car")
     public Car getCar(){
@@ -14,6 +20,6 @@ public class CarController {
 
     @GetMapping("/greet")
     public String greet(){
-        return "Hello, World!";
+        return carService.greet();
     }
 }
