+++ src/systemTest/java/com/example/DevOpsExaminationProject/CarSystemTest.java
@@ -1,25 +1,21 @@
+package com.example.DevOpsExaminationProject;
+
 import io.restassured.RestAssured;
 
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
 
 import static io.restassured.RestAssured.given;
+import static io.restassured.RestAssured.port;
 import static org.hamcrest.Matchers.equalTo;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class GreetingSystemTest {
 
-
-    @LocalServerPort
-    private int port;
+class CarSystemTest {
 
     @BeforeEach
     void setUp(){
         RestAssured.baseURI = "http://localhost";
-        RestAssured.port = port;
+        port = port;
     }
 
     @Test
