--- src/systemTest/java/controllers/CarControllerTest.java
@@ -1,35 +1,0 @@
-package controllers;
-
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.notNullValue;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class CarControllerTest {
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    void setUp(){
-        RestAssured.baseURI = "http://localhost";
-        RestAssured.port = port;
-    }
-
-    @Test
-    void getCar() {
-        given()
-                .when()
-                .get("/car")
-                .then()
-                .statusCode(200)
-                .assertThat()
-                .body("light", notNullValue())
-                .body("engine", notNullValue());
-    }
-}
